Move http endpoints to `service` module
