+++ service/src/main/scala/org/fiume/sketch/http/DocumentsRoutes.scala
@@ -1,4 +1,4 @@
-package org.fiume.sketch.storage.documents.http
+package org.fiume.sketch.http
 
 import cats.MonadThrow
 import cats.data.EitherT
@@ -9,6 +9,17 @@ import io.circe.{Decoder, Encoder, HCursor, *}
 import io.circe.{Json as JJson}
 import io.circe.Decoder.Result
 import io.circe.syntax.*
+import org.fiume.sketch.http.DocumentsRoutes.{
+  AuthorQueryParamMatcher,
+  DocumentIdVar,
+  Line,
+  Linebreak,
+  NewlineDelimitedJson,
+  NewlineDelimitedJsonEncoder,
+  OwnerQueryParamMatcher
+}
+import org.fiume.sketch.http.DocumentsRoutes.Model.*
+import org.fiume.sketch.http.DocumentsRoutes.Model.Json.given
 import org.fiume.sketch.shared.app.EntityId.given
 import org.fiume.sketch.shared.app.http4s.middlewares.SemanticInputError
 import org.fiume.sketch.shared.app.troubleshooting.ErrorDetails
@@ -21,17 +32,6 @@ import org.fiume.sketch.shared.domain.documents.Document.Metadata.*
 import org.fiume.sketch.shared.domain.documents.algebras.DocumentsStore
 import org.fiume.sketch.shared.typeclasses.FromStringSyntax.*
 import org.fiume.sketch.shared.typeclasses.SemanticStringSyntax.*
-import org.fiume.sketch.storage.documents.http.DocumentsRoutes.{
-  AuthorQueryParamMatcher,
-  DocumentIdVar,
-  Line,
-  Linebreak,
-  NewlineDelimitedJson,
-  NewlineDelimitedJsonEncoder,
-  OwnerQueryParamMatcher
-}
-import org.fiume.sketch.storage.documents.http.DocumentsRoutes.Model.*
-import org.fiume.sketch.storage.documents.http.DocumentsRoutes.Model.Json.given
 import org.http4s.{Charset, EntityEncoder, HttpRoutes, MediaType, *}
 import org.http4s.MediaType.application
 import org.http4s.circe.CirceEntityDecoder.*
