+++ service/src/test/resources/contract/documents/document.metadata.request.json
