+++ service/src/test/resources/contract/documents/document.uuid.response.json
